Split chat route tests
